webmap/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
pillow==12.0.0
numpy==2.2.6
aiohttp==3.13.2
aiofiles==25.1.0
pydantic==2.12.3
jinja2==3.1.6